+++ RazorSvelte/App/login.svelte
@@ -5,20 +5,20 @@
 
 <Layout>
     <div class="container text-center m-5">
-        <ul class="list-group">
-            <li class="list-group-item">
+        <ul class="list-group border-0">
+            <li class="list-group-item border-0">
                 <a class="btn btn-primary btn-lg" href="{urls.signInGoogleUrl}">
                     <i class="bi bi-google"></i>
                     Login With Google
                 </a>
             </li>
-            <li class="list-group-item">
+            <li class="list-group-item border-0">
                 <a class="btn btn-primary btn-lg" href="{urls.signInLinkedInUrl}">
                     <i class="bi bi-linkedin"></i>
                     Login With LinkedIn
                 </a>
             </li>
-            <li class="list-group-item">
+            <li class="list-group-item border-0">
                 <a class="btn btn-primary btn-lg" href="{urls.signInGitHubUrl}">
                     <i class="bi bi-github"></i>
                     Login With GitHub